Project refactoring. Renaming
