--- Errors/file-access-error.js
@@ -1,7 +1,0 @@
-module.exports =
-  class FileAccessError extends Error {
-    constructor(message) {
-      super(message);
-      this.name = "FileAccessError";
-    }
-  };
