--- Errors/validation-error.js
@@ -1,7 +1,0 @@
-module.exports =
-  class ValidationError extends Error {
-    constructor(message) {
-      super(message);
-      this.name = "ValidationError";
-    }
-  };
