+++ errors/file-access-error.js
@@ -1,0 +1,8 @@
+class FileAccessError extends Error {
+  constructor(message) {
+    super(message);
+    this.name = "FileAccessError";
+  }
+}
+
+module.exports = FileAccessError;
