--- helpers.js
@@ -1,3 +1,0 @@
-exports.isEnglishLetter = (char) => {
-  return char && char.length === 1 && /[a-z]/i.test(char);
-}
